# app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "PMRP Project ",
        "description": "Application Programming Interface (API)"
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "paths": {
...
        "/static/{code}": {
            "get": {
                "tags": [
                    "Static"
                ],
                "description": "",
                "parameters": [
                    {
                        "in": "path",
                        "name": "code",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File"
                    }
                }
            }
        },
...
        "/upload/document": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "properties": {
                                "typeFile": {
                                    "type": "string",
                                    "example": "company"
                                }
                            }
                        }
                    },
                    {
                        "in": "formData",
                        "name": "Profile",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "description": "files upload multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "File is uploaded"
                                },
                                "data": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/profile": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "properties": {
                                "typeFile": {
                                    "type": "string",
                                    "example": "company"
                                }
                            }
                        }
                    },
                    {
                        "in": "formData",
                        "name": "Profile",
                        "type": "file",
                        "description": "file upload"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "File is uploaded"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "orgin_name": {
                                            "type": "string",
                                            "example": "file_1.png"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "ca9abdf8-d3eb-41f2-8a84-a5c3a8cd83e1_file_1.jpg"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "Li91cGxvYWRzL2NvbXBhbnkvY2E5YWJkZjgtZDNlYi00MWY"
                                        },
                                        "mimetype": {
                                            "type": "string",
                                            "example": "image/jpeg"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 285384
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/excel": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "properties": {
                                "typeFile": {
                                    "type": "string",
                                    "example": "Excel"
                                }
                            }
                        }
                    },
                    {
                        "in": "formData",
                        "name": "excel",
                        "type": "file",
                        "description": "file upload"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "File is uploaded"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "orgin_name": {
                                            "type": "string",
                                            "example": "file_1.xlsx"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "ca9abdf8-d3eb-41f2-8a84-a5c3a8cd83e1_file_1.xlsx"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "Li91cGxvYWRzL2NvbXBhbnkvY2E5YWJkZjgtZDNlYi00MWY"
                                        },
                                        "mimetype": {
                                            "type": "string",
                                            "example": "xls/xlsx"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 285384
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}